refactor (step5) : Winners 추가, RacingManager Cars 매개변수로 수정
