+++ src/main/kotlin/racingcar/car/Winners.kt
@@ -1,0 +1,4 @@
+package racingcar.car
+
+@JvmInline
+value class Winners(val winners: List<Car>)
